refactor: moved src/lib to src/utils
