+++ src/utils/removeComments.ts
